Move get_blob_as_stream to util file
