+++ tests/functions/test_visualizations.py
@@ -11,7 +11,6 @@ from functions.visualizations import (
     DownloadableFiles,
     _ClustergrammerTransform,
     _npx_to_dataframe,
-    _get_blob_as_stream,
 )
 
 from tests.util import make_pubsub_event
@@ -58,7 +57,7 @@ def test_npx_clustergrammer_end_to_end(monkeypatch, metadata_df):
     fake_npx = open(NPX_PATH, "rb")
     _get_blob_as_stream.return_value = fake_npx
     monkeypatch.setattr(
-        functions.visualizations, "_get_blob_as_stream", _get_blob_as_stream
+        functions.visualizations, "get_blob_as_stream", _get_blob_as_stream
     )
 
     # Mock metadata_df
@@ -112,24 +111,3 @@ def test_npx_to_dataframe():
         {"CTTTTPPS1.01": [1, -1], "CTTTTPPS2.01": [-1, 1]}, index=["Assay1", "Assay2"]
     )
     assert expected_df.equals(npx_df)
-
-
-def test_get_blob_as_stream(monkeypatch):
-    """Ensure GCS blob utility works as expected"""
-    blob_str = "foo bar"
-    blob_bytes = blob_str.encode()
-    get_blob_bytes = MagicMock()
-    get_blob_bytes.return_value = blob_bytes
-    monkeypatch.setattr(
-        functions.visualizations, "__download_blob_bytes", get_blob_bytes
-    )
-
-    # as BytesIO
-    stream = _get_blob_as_stream("")
-    assert isinstance(stream, BytesIO)
-    assert stream.read() == blob_bytes
-
-    # as StringIO
-    stream = _get_blob_as_stream("", as_string=True)
-    assert isinstance(stream, StringIO)
-    assert stream.read() == blob_str
